polyval: rename all backends to *_backend
It might make sense to refactor these into a containing `backend` module
as well, but for now these are the only additional modules in the crate
besides `lib.rs`, so just label them as backends.
